tast: change software dependency for arc on reven
Despite enabling the ARC on the reven board, the
Android image installation is initiated in the Chrome
browser (with DLC method) and is subject to hardware
specification restrictions. Consequently, not all reven
devices will have the Android image. Initially, manual
testing is scheduled, with potential adjustments to the
test in the future if necessary.

BUG=b:348453057
TEST=skip the arc tests on reven* board

Change-Id: I519e936f6f4710e8377a8a3711c5c559f042aa0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/6111443
Auto-Submit: Jiao Zhou <jiaozhou@google.com>
Commit-Queue: Jiao Zhou <jiaozhou@google.com>
Reviewed-by: Muhammad Hasan Khan <mhasank@chromium.org>
Tested-by: Jiao Zhou <jiaozhou@google.com>