ish: Add hwdep.ChromeISH condition
This condition will allow us to filter ISH tests at runtime. Note that
the test run as a part of this CL is not yet uploaded, but does include
the line:

  HardwareDeps: hwdep.D(hwdep.ChromeEC(), hwdep.ChromeISH()),

Add the dsp group of tests and a sub-attribute for ISH
specific tests.

BUG=b:401019287
TEST=tast run $DUT_IP firmware.RebootAndCheckIshmod
TEST=tast list '("group:dsp" && "dsp_ish")'

Change-Id: I864d7541ff872fccd386df76ec61aaf28e64be05

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/6329965
Reviewed-by: Seewai Fu <seewaifu@google.com>
Auto-Submit: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>