tast: Fix arch check condition in IsIntelUarchOlderThan()
The IsIntelUarchOlderThan() hwdep routine is supposed to identify if
the DUT core is within the provided IntelBigCoreOrderList. However, an
missing equality condition check led to the most recent platform in the
list being incorrectly excluded. Consequently, the testcase
CheckEOPState.get_eop_state was skipped for the latest Intel platform.
This patch updates the necessary conditional statement to rectify this.

BUG=b:403491449
TEST=testcase CheckEOPState.get_eop_state doesn't skip for MTL and PTL.
TEST=platform/tast/fast_build.sh -T

Change-Id: I7ea8599eedb697b93e7e552cf444725aef235463
Signed-off-by: Dinesh Gehlot <digehlot@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/6406514
Reviewed-by: Seewai Fu <seewaifu@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>